requirements.txt
fastapi
uvicorn
pydantic
tensorflow-object-detection-api
